concatenate/requirements.txt
websocket-client
django
pyenchant